requirements.txt
groq
streamlit
python-dotenv